Funcionamiento de todos los JRadioButton y funcionamiento correcto de un reporte falta la logica de los otros reportes y detalles

+++ src/Game_1/Swimmer.java
@@ -41,7 +41,7 @@ public class Swimmer {
         return time;
     }
 
-    public void setTime(int time) {
+    public void setTime(long time) {
         this.time = time;
     }
